- Turned on warning level 4 just to see what it would produce: a lot of   warnings. At first, I was going to try and clean them all up. Then I decided   that was a worthless cause and went about just acting on the ones that   might actually be helpful:    C4189 (local variable is initialized but not referenced)    C4702 (unreachable code)    C4512 (assignment operator could not be generated)
SVN r420 (trunk)

+++ src/g_strife/a_programmer.cpp
@@ -207,7 +207,7 @@ void AProgLevelEnder::Tick ()
 
 PalEntry AProgLevelEnder::GetBlend ()
 {
-	return PalEntry (special1, 0, 0, 0);
+	return PalEntry ((BYTE)special1, 0, 0, 0);
 }
 
 //============================================================================
